requirements.txt
discord.py
glitch-this
